Move store to separate file
